Clean paths and reformat for current repo
This commit cleans paths that point to locations on minerva and replaces them with more general paths. This commit also reformats my workflow/finemapping_multi_GWAS_withLD_noLD_autoVEP_paralleljobs file so that it uses the --use-conda flag and works with the SAFFARI repo as it currently is set up.